Implement basic prototype of automated cube gripping in mobility package
The targetHandler callback in mobility has been updated to execute a tag alignment procedure, followed by an actuation procedure that prepares the gripper for engagement with the cube. The state machine has also been modified to include a simple gripping procedure to pick up and transport the cube, as well as a cube drop-off procedure that deposits cubes in the center collection zone. This prototype will require experimental tuning to ensure predictable, accurate pick up and drop off of cubes.